% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore.R
\name{.create.gitignore}
\alias{.create.gitignore}
\title{Create a .gitignore file tailored that meets usual Sandpiper requirements}
\usage{
.create.gitignore(filepath = "./.gitignore")
}
\arguments{
\item{filepath}{Character string specifying the path and name of the
.gitignore file to be created. Defaults to ".gitignore" (in the current
working directory).}
}
\value{
This function does not return any value. It creates a file in the
file system as a side effect.
}
\description{
This function generates a .gitignore file that is commonly used for sandpiper analytics projects.
It initially ignores all files but then selectively un-ignores specific file
types essential for R development, such as R scripts, R Markdown files,
R project files, and others. It also allows all subdirectories.
}
\examples{
# Create .gitignore in the current working directory
create_gitignore()

# Create .gitignore in a specific directory
create_gitignore("path/to/your/project/.gitignore")

}
