% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitignore.R
\name{.create.renvignore}
\alias{.create.renvignore}
\title{Create a .renvignore to aid with dependency detection}
\usage{
.create.renvignore(filepath = "./.renvignore")
}
\arguments{
\item{filepath}{Character string specifying the path and name of the
.renvignore file to be created. Defaults to ".renvignore" (in the current
working directory).}
}
\value{
This function does not return any value. It creates a file in the
file system as a side effect.
}
\description{
This function generates a .renvignore that makes sure dependencies not
already declared in DESCRIPTION can be detected.
}
\examples{
# Create .gitignore in the current working directory
.create.renvignore()

# Create .gitignore in a specific directory
.create.renvignore("path/to/your/project/.gitignore")

}
